pyproject.toml
[tool.poetry]
name = "llm-chatbot"
version = "0.1.0"
description = "A simple chatbot application that allows you to interact with an LLM model"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
python-dotenv = "^1.0.0"
openai = "^1.3.0"
pydantic = "^2.4.2"
pydantic-settings = "^2.0.3"
jinja2 = "^3.1.2"
httpx = "^0.25.0"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.0"
black = "^23.7.0"
isort = "^5.12.0"
mypy = "^1.5.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88

[tool.isort]
profile = "black" 